[#14] feat: revviewImageFile 매핑 테스트
